+++ packages/back-top/TyhBackTop.vue
@@ -12,17 +12,16 @@
 <script setup>
 import { ref, onMounted } from 'vue'
 defineProps({
-  bottom: { // 距离下面的定位距离
+  bottom: {
     type: String,
     default: '50'
   },
-  right: { // 距离右面的定位距离
+  right: {
     type: String,
     default: '50'
   }
 })
 const backShow = ref(false)
-// 监视按钮显示状态
 function handleScroll () {
   let timer = null
   return function () {
@@ -35,7 +34,6 @@ function handleScroll () {
     }, 188)
   }
 }
-// 点击按钮回到顶部
 function backTop () {
   let top = document.documentElement.scrollTop || document.body.scrollTop
   const timeTop = setInterval(() => {